Refactor ConfirmModal
Refactor DocumentationInfoModal

Refactor InfoConfirmModal

Refactor InfoModal
