+++ app/components/modals/DocumentationInfoModal.jsx
@@ -1,0 +1,14 @@
+import InfoModal from "./InfoModal";
+import { Documentation } from "shared";
+import style from "./Modals.module.css";
+
+const DocumentationInfoModal = ({ document, ...props }) => (
+  <InfoModal
+    {...props}
+    modalContent={
+      <Documentation name={document} className={style.infoModalDocumentation} />
+    }
+  />
+);
+
+export default DocumentationInfoModal;
