--- app/components/modals/InfoDocumentModal/index.js
@@ -1,11 +1,0 @@
-import InfoModal from "../InfoModal/Modal";
-import { Documentation } from "shared";
-
-export default ({ document, ...props }) => (
-  <InfoModal
-    {...props}
-    modalContent={
-      <Documentation name={document} className="info-modal-documentation" />
-    }
-  />
-);
