+++ app/components/modals/index.js
@@ -1,9 +1,9 @@
 export { default as ImportScriptModal } from "./ImportScriptModal/ImportScriptModal";
 export { default as AddAccountModal } from "./AddAccountModal/AddAccountModal";
 export { default as ChangePassphraseModal } from "./ChangePassphraseModal/ChangePassphraseModal";
-export { default as InfoModal } from "./InfoModal/Modal";
+export { default as InfoModal } from "./InfoModal";
 export { default as InfoConfirmModal } from "./InfoConfirmModal";
-export { default as InfoDocumentModal } from "./InfoDocumentModal";
+export { default as DocumentationInfoModal } from "./DocumentationInfoModal";
 export { default as PassphraseModal } from "./PassphraseModal/PassphraseModal";
 export { default as ConfirmModal } from "./ConfirmModal";
 export { default as SeedCopyConfirmModal } from "./SeedCopyConfirmModal/SeedCopyConfirmModal";
